+++ src/reducers/controlReducer.ts
@@ -1,6 +1,6 @@
 import * as ol from 'openlayers';
 
-import { findDataset, findDatasetVariable, getDatasetTimeRange, findDatasetOrUserPlace } from '../model/dataset';
+import { findDataset, findDatasetVariable, getDatasetTimeRange } from '../model/dataset';
 import { ControlState, newControlState } from '../states/controlState';
 import {
     SELECT_DATASET,
@@ -18,7 +18,7 @@ import {
     OPEN_DIALOG,
     CLOSE_DIALOG,
     INC_SELECTED_TIME,
-    UPDATE_SETTINGS,
+    UPDATE_SETTINGS, SET_MAP_INTERACTION,
 } from '../actions/controlActions';
 import { CONFIGURE_SERVERS, DataAction, ADD_USER_PLACE, REMOVE_USER_PLACE } from "../actions/dataActions";
 import { I18N } from "../config";
@@ -73,9 +73,9 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
         case SELECT_PLACE: {
             const selectedPlaceId = action.selectedPlaceId;
             let flyTo = state.flyTo;
-            if (selectedPlaceId) {
-                const place = findDatasetOrUserPlace(action.datasets, action.userPlaceGroup, selectedPlaceId);
-                if (place !== null) {
+            if (selectedPlaceId && action.showInMap) {
+                const place = action.places.find(p => p.id === selectedPlaceId);
+                if (place) {
                     if (place.bbox && place.bbox.length === 4) {
                         flyTo = place.bbox as [number, number, number, number];
                     } else if (place.geometry && SIMPLE_GEOMETRY_TYPES.includes(place.geometry.type)) {
@@ -194,6 +194,12 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
                 };
             }
             return state;
+        }
+        case SET_MAP_INTERACTION: {
+            return {
+                ...state,
+                mapInteraction: action.mapInteraction
+            };
         }
         case ADD_ACTIVITY: {
             return {